Try to handle dropping of nubs which are not registered as actors/commanders/acceptors.
This showed up after a bug in hubCommands/setUsername() did a
dropCommander() but not addCommander(). OK, that was lazy logic, but
at some level acceptably clean. Everything was then OK, until the
commander logged out, and the hub started spinning.

The way that nubs get dropped on a disconnect is for the IO layer
(what the select() loop calls to call nub.shutdown(). That needs to
eventually call nub.ioshutdown(), which actually removes the object
from the select table. Most nubs do this by calling a highlevel
routine (hub.dropNub()), which unregisters the nub from the management
tables then calls back to the io.shutdown(). In this case the nub no
longer had a top-level personality, and the io.shutdown() never got called.

This change tries to call nub.shutdown() in the case when no hub-level
nub is found.